Add handling for setting number of packs
Add a method to the Game class to allow the number of packs to use in the deck to be set. Also add testing for the new method.

Also change the game flow slightly, so that the number of starting chips is asked for before the number of players is asked for.